+++ test/rljson_test.dart
@@ -10,14 +10,32 @@ import 'package:test/test.dart';
 void main() {
   group('Rljson', () {
     final rljson = Rljson.example;
+    late String a0Hash;
+    late String a1Hash;
+    late String b0Hash;
+    late String b1Hash;
+
+    setUp(
+      () {
+        a0Hash = rljson.hash(table: '@tableA', index: 0);
+        a1Hash = rljson.hash(table: '@tableA', index: 1);
+        b0Hash = rljson.hash(table: '@tableB', index: 0);
+        b1Hash = rljson.hash(table: '@tableB', index: 1);
+
+        expect(a0Hash.length, 22);
+        expect(a1Hash.length, 22);
+        expect(b0Hash.length, 22);
+        expect(b1Hash.length, 22);
+      },
+    );
 
     group('ls()', () {
       test('lists the pathes of all items', () {
         expect(rljson.ls(), [
-          '@layerA/KFQrf4mEz0UPmUaFHwH4T6/keyA0',
-          '@layerA/YPw-pxhqaUOWRFGramr4B1/keyA1',
-          '@layerB/nmejjLAUhygiT6WFDPPsHy/keyB0',
-          '@layerB/dXhIygNwNMVPEqFbsFJkn6/keyB1',
+          '@tableA/KFQrf4mEz0UPmUaFHwH4T6/keyA0',
+          '@tableA/YPw-pxhqaUOWRFGramr4B1/keyA1',
+          '@tableB/nmejjLAUhygiT6WFDPPsHy/keyB0',
+          '@tableB/dXhIygNwNMVPEqFbsFJkn6/keyB1',
         ]);
       });
     });
@@ -25,106 +43,106 @@ void main() {
     group('fromData(data)', () {
       test('adds hashes to all fields', () {
         expect(rljson.data, {
-          '@layerA': {
-            'KFQrf4mEz0UPmUaFHwH4T6': {
+          '@tableA': {
+            a0Hash: {
               'keyA0': 'a0',
-              '_hash': 'KFQrf4mEz0UPmUaFHwH4T6',
+              '_hash': a0Hash,
             },
-            'YPw-pxhqaUOWRFGramr4B1': {
+            a1Hash: {
               'keyA1': 'a1',
-              '_hash': 'YPw-pxhqaUOWRFGramr4B1',
+              '_hash': a1Hash,
             },
           },
-          '@layerB': {
-            'nmejjLAUhygiT6WFDPPsHy': {
+          '@tableB': {
+            b0Hash: {
               'keyB0': 'b0',
-              '_hash': 'nmejjLAUhygiT6WFDPPsHy',
+              '_hash': b0Hash,
             },
-            'dXhIygNwNMVPEqFbsFJkn6': {
+            b1Hash: {
               'keyB1': 'b1',
-              '_hash': 'dXhIygNwNMVPEqFbsFJkn6',
+              '_hash': b1Hash,
             },
           },
         });
       });
     });
 
-    group('layer(String layer)', () {
+    group('table(String table)', () {
       group('returns', () {
-        test('the layer when existing', () {
-          final layer = rljson.layer('@layerA');
-          expect(layer, {
-            'KFQrf4mEz0UPmUaFHwH4T6': {
+        test('the table when existing', () {
+          final table = rljson.table('@tableA');
+          expect(table, {
+            a0Hash: {
               'keyA0': 'a0',
-              '_hash': 'KFQrf4mEz0UPmUaFHwH4T6',
+              '_hash': a0Hash,
             },
-            'YPw-pxhqaUOWRFGramr4B1': {
+            a1Hash: {
               'keyA1': 'a1',
-              '_hash': 'YPw-pxhqaUOWRFGramr4B1',
+              '_hash': a1Hash,
             },
           });
         });
       });
 
       group('throws', () {
-        test('when layer does not exist', () {
+        test('when table does not exist', () {
           late final Exception exception;
 
           try {
-            rljson.layer(
-              '@layerC',
+            rljson.table(
+              '@tableC',
             );
           } catch (e) {
             exception = e as Exception;
           }
 
-          expect(exception.toString(), 'Exception: Layer not found: @layerC');
+          expect(exception.toString(), 'Exception: Table not found: @tableC');
         });
       });
     });
 
-    group('items(layer, where)', () {
+    group('items(table, where)', () {
       test('returns the items that match the query', () {
         final items = rljson.items(
-          layer: '@layerA',
+          table: '@tableA',
           where: (item) => item['keyA0'] == 'a0',
         );
 
         expect(items, [
-          {'keyA0': 'a0', '_hash': 'KFQrf4mEz0UPmUaFHwH4T6'},
+          {'keyA0': 'a0', '_hash': a0Hash},
         ]);
       });
     });
 
-    group('item(layer, hash)', () {
+    group('item(table, hash)', () {
       group('returns', () {
         test('the item when existing', () {
-          final item = rljson.item('@layerA', 'KFQrf4mEz0UPmUaFHwH4T6');
+          final item = rljson.item('@tableA', a0Hash);
           expect(item, {
             'keyA0': 'a0',
-            '_hash': 'KFQrf4mEz0UPmUaFHwH4T6',
+            '_hash': a0Hash,
           });
         });
       });
 
       group('throws', () {
-        test('when layer is not available', () {
+        test('when table is not available', () {
           late final Exception exception;
 
           try {
-            rljson.item('@layerC', 'KFQrf4mEz0UPmUaFHwH4T6');
+            rljson.item('@tableC', a0Hash);
           } catch (e) {
             exception = e as Exception;
           }
 
-          expect(exception.toString(), 'Exception: Layer not found: @layerC');
+          expect(exception.toString(), 'Exception: Table not found: @tableC');
         });
 
         test('when hash is not available', () {
           late final Exception exception;
 
           try {
-            rljson.item('@layerA', 'nonExistingHash');
+            rljson.item('@tableA', 'nonExistingHash');
           } catch (e) {
             exception = e as Exception;
           }
@@ -132,20 +150,20 @@ void main() {
           expect(
             exception.toString(),
             'Exception: Item not found with hash "nonExistingHash" '
-            'in layer "@layerA"',
+            'in table "@tableA"',
           );
         });
       });
     });
 
-    group('value(layer, hash, key, key2, key3, key4 followLinks)', () {
+    group('value(table, hash, key, key2, key3, key4 followLinks)', () {
       group('returns', () {
-        test('the value of the key of the item with hash in layer', () {
+        test('the value of the key of the item with hash in table', () {
           expect(
-            rljson.value(
-              layer: '@layerA',
-              itemHash: 'KFQrf4mEz0UPmUaFHwH4T6',
-              key: 'keyA0',
+            rljson.get(
+              table: '@tableA',
+              item: a0Hash,
+              key1: 'keyA0',
             ),
             'a0',
           );
@@ -153,34 +171,36 @@ void main() {
 
         test('the complete item, when no key is given', () {
           expect(
-            rljson.value(layer: '@layerA', itemHash: 'KFQrf4mEz0UPmUaFHwH4T6'),
-            {'keyA0': 'a0', '_hash': 'KFQrf4mEz0UPmUaFHwH4T6'},
+            rljson.get(table: '@tableA', item: a0Hash),
+            {'keyA0': 'a0', '_hash': a0Hash},
           );
         });
 
         test('the linked value, when property is a link', () {
           final rljson = Rljson.exampleWithLink;
 
+          final tableALinkHash = rljson.hash(table: '@linkToTableA', index: 0);
+
           expect(
-            rljson.value(
-              layer: '@linkToLayerA',
-              itemHash: 'Cr5kvgDz5NGgbnuHu-Z05N',
-              key: '@layerA',
+            rljson.get(
+              table: '@linkToTableA',
+              item: tableALinkHash,
+              key1: '@tableA',
             ),
-            {'_hash': 'KFQrf4mEz0UPmUaFHwH4T6', 'keyA0': 'a0'},
+            {'_hash': a0Hash, 'keyA0': 'a0'},
           );
         });
 
-        test('the linked value accross multiple layers using key2 to key4', () {
+        test('the linked value accross multiple tables using key2 to key4', () {
           final rljson = Rljson.exampleWithDeepLink;
           // print(const JsonEncoder.withIndent(' ')
           // .convert(rljson.originalData));
           // return;
           expect(
-            rljson.value(
-              layer: '@a',
-              itemHash: 'GnSVp1CmoAo3rPiiGl44p-',
-              key: '@b',
+            rljson.get(
+              table: '@a',
+              item: 'GnSVp1CmoAo3rPiiGl44p-',
+              key1: '@b',
               key2: '@c',
               key3: '@d',
               key4: 'value',
@@ -195,10 +215,10 @@ void main() {
           late final Exception exception;
 
           try {
-            rljson.value(
-              layer: '@layerA',
-              itemHash: 'KFQrf4mEz0UPmUaFHwH4T6',
-              key: 'nonExistingKey',
+            rljson.get(
+              table: '@tableA',
+              item: a0Hash,
+              key1: 'nonExistingKey',
             );
           } catch (e) {
             exception = e as Exception;
@@ -207,7 +227,7 @@ void main() {
           expect(
             exception.toString(),
             'Exception: Key "nonExistingKey" not found in item with hash '
-            '"KFQrf4mEz0UPmUaFHwH4T6" in layer "@layerA"',
+            '"KFQrf4mEz0UPmUaFHwH4T6" in table "@tableA"',
           );
         });
 
@@ -215,10 +235,10 @@ void main() {
           late final Exception exception;
 
           try {
-            rljson.value(
-              layer: '@layerA',
-              itemHash: 'KFQrf4mEz0UPmUaFHwH4T6',
-              key: 'keyA0',
+            rljson.get(
+              table: '@tableA',
+              item: a0Hash,
+              key1: 'keyA0',
               key2: 'keyA1',
             );
           } catch (e) {
@@ -237,14 +257,37 @@ void main() {
 
     group('addData(data)', () {
       group('throws', () {
-        test('when layer names do not start with @', () {
+        test('when validateHashes is true and hashes are missing', () {
+          late final Exception exception;
+
+          try {
+            rljson.addData(
+              {
+                '@tableA': {
+                  '_data': [
+                    {'keyA0': 'a0'},
+                  ],
+                },
+              },
+              validateHashes: true,
+            );
+          } catch (e) {
+            exception = e as Exception;
+          }
+
+          expect(
+            exception.toString(),
+            'Exception: Hash is missing.',
+          );
+        });
+        test('when table names do not start with @', () {
           late final Exception exception;
 
           try {
             rljson.addData({
-              'layerA': {'_data': <dynamic>[]},
-              'layerB': {'_data': <dynamic>[]},
-              'layerC': {'_data': <dynamic>[]},
+              'tableA': {'_data': <dynamic>[]},
+              'tableB': {'_data': <dynamic>[]},
+              'tableC': {'_data': <dynamic>[]},
             });
           } catch (e) {
             exception = e as Exception;
@@ -252,17 +295,17 @@ void main() {
 
           expect(
             exception.toString(),
-            'Exception: Layer name must start with @: layerA',
+            'Exception: Table name must start with @: tableA',
           );
         });
 
-        test('when layers do not contain a _data object', () {
+        test('when tables do not contain a _data object', () {
           late final Exception exception;
 
           try {
             rljson.addData({
-              '@layerA': <String, dynamic>{},
-              '@layerB': <String, dynamic>{},
+              '@tableA': <String, dynamic>{},
+              '@tableB': <String, dynamic>{},
             });
           } catch (e) {
             exception = e as Exception;
@@ -270,19 +313,19 @@ void main() {
 
           expect(
             exception.toString(),
-            'Exception: _data is missing in layer: @layerA, @layerB',
+            'Exception: _data is missing in table: @tableA, @tableB',
           );
         });
 
-        test('when layers do not contain a _data that is not a list', () {
+        test('when tables do not contain a _data that is not a list', () {
           late final Exception exception;
 
           try {
             rljson.addData({
-              '@layerA': <String, dynamic>{
+              '@tableA': <String, dynamic>{
                 '_data': <dynamic>{},
               },
-              '@layerB': <String, dynamic>{
+              '@tableB': <String, dynamic>{
                 '_data': <dynamic>{},
               },
             });
@@ -293,31 +336,31 @@ void main() {
           expect(
             exception.toString(),
             'Exception: '
-            '_data must be a list in layer: @layerA, @layerB',
+            '_data must be a list in table: @tableA, @tableB',
           );
         });
       });
 
       test('adds data to the json', () {
         final rljson2 = rljson.addData({
-          '@layerA': {
+          '@tableA': {
             '_data': [
               {'keyA2': 'a2'},
             ],
           },
         });
 
-        final items = rljson2.originalData['@layerA']['_data'] as List<dynamic>;
+        final items = rljson2.originalData['@tableA']['_data'] as List<dynamic>;
         expect(items, [
-          {'keyA0': 'a0', '_hash': 'KFQrf4mEz0UPmUaFHwH4T6'},
-          {'keyA1': 'a1', '_hash': 'YPw-pxhqaUOWRFGramr4B1'},
+          {'keyA0': 'a0', '_hash': a0Hash},
+          {'keyA1': 'a1', '_hash': a1Hash},
           {'keyA2': 'a2', '_hash': 'apLP3I2XLnVm13umIZdVhV'},
         ]);
       });
 
-      test('replaces data when the added layer is not yet existing', () {
+      test('replaces data when the added table is not yet existing', () {
         final rljson2 = rljson.addData({
-          '@layerC': {
+          '@tableC': {
             '_data': [
               {'keyC0': 'c0'},
             ],
@@ -326,27 +369,27 @@ void main() {
 
         final items = rljson2.ls();
         expect(items, [
-          '@layerA/KFQrf4mEz0UPmUaFHwH4T6/keyA0',
-          '@layerA/YPw-pxhqaUOWRFGramr4B1/keyA1',
-          '@layerB/nmejjLAUhygiT6WFDPPsHy/keyB0',
-          '@layerB/dXhIygNwNMVPEqFbsFJkn6/keyB1',
-          '@layerC/afNjjrfH8-OfkkEH1uCK14/keyC0',
+          '@tableA/KFQrf4mEz0UPmUaFHwH4T6/keyA0',
+          '@tableA/YPw-pxhqaUOWRFGramr4B1/keyA1',
+          '@tableB/nmejjLAUhygiT6WFDPPsHy/keyB0',
+          '@tableB/dXhIygNwNMVPEqFbsFJkn6/keyB1',
+          '@tableC/afNjjrfH8-OfkkEH1uCK14/keyC0',
         ]);
       });
 
       test('does not cause duplicates', () {
         final rljson2 = rljson.addData({
-          '@layerA': {
+          '@tableA': {
             '_data': [
               {'keyA1': 'a1'},
             ],
           },
         });
 
-        final items = rljson2.originalData['@layerA']['_data'] as List<dynamic>;
+        final items = rljson2.originalData['@tableA']['_data'] as List<dynamic>;
         expect(items, [
-          {'keyA0': 'a0', '_hash': 'KFQrf4mEz0UPmUaFHwH4T6'},
-          {'keyA1': 'a1', '_hash': 'YPw-pxhqaUOWRFGramr4B1'},
+          {'keyA0': 'a0', '_hash': a0Hash},
+          {'keyA1': 'a1', '_hash': a1Hash},
         ]);
       });
     });
@@ -358,19 +401,21 @@ void main() {
       });
 
       group('throws', () {
-        test('when the layer of a link does not exist', () {
+        test('when the table of a link does not exist', () {
           final rljson = Rljson.exampleWithLink;
 
-          // Add an item with an link to a non-existing layer
+          // Add an item with an link to a non-existing table
           final jsonWithBrokenLink = rljson.addData({
-            '@layerA': {
+            '@tableA': {
               '_data': [
                 {
-                  '@nonExistingLayer': 'a2',
+                  '@nonExistingTable': 'a2',
                 },
               ],
             },
           });
+
+          final a0Hash = jsonWithBrokenLink.hash(table: '@tableA', index: 2);
 
           late final String message;
 
@@ -382,24 +427,29 @@ void main() {
 
           expect(
             message,
-            'Exception: Layer "@layerA" has an item "isQfTSg24p0hXHxkBB_wEa" '
-            'which links to not existing layer "@nonExistingLayer".',
+            'Exception: Table "@tableA" has an item "$a0Hash" '
+            'which links to not existing table "@nonExistingTable".',
           );
         });
 
         test('when linked item does not exist', () {
           final rljson = Rljson.exampleWithLink;
 
-          // Add an item with an link to a non-existing layer
+          // Add an item with an link to a non-existing table
           final jsonWithBrokenLink = rljson.addData({
-            '@linkToLayerA': {
+            '@linkToTableA': {
               '_data': [
                 {
-                  '@layerA': 'brokenHash',
+                  '@tableA': 'brokenHash',
                 },
               ],
             },
           });
+
+          final linkToTableAHash = jsonWithBrokenLink.hash(
+            table: '@linkToTableA',
+            index: 1,
+          );
 
           late final String message;
 
@@ -411,9 +461,9 @@ void main() {
 
           expect(
             message,
-            'Exception: Layer "@linkToLayerA" has an item '
-            '"NnQGoODqzFIwANtgDUMkhA" which links to not existing '
-            'item "brokenHash" in layer "@layerA".',
+            'Exception: Table "@linkToTableA" has an item '
+            '"$linkToTableAHash" which links to not existing '
+            'item "brokenHash" in table "@tableA".',
           );
         });
       });
@@ -423,24 +473,24 @@ void main() {
       group('returns the data where the _data list is replaced by a map', () {
         test('with example', () {
           expect(rljson.data, <String, dynamic>{
-            '@layerA': {
-              'KFQrf4mEz0UPmUaFHwH4T6': {
+            '@tableA': {
+              a0Hash: {
                 'keyA0': 'a0',
-                '_hash': 'KFQrf4mEz0UPmUaFHwH4T6',
+                '_hash': a0Hash,
               },
-              'YPw-pxhqaUOWRFGramr4B1': {
+              a1Hash: {
                 'keyA1': 'a1',
-                '_hash': 'YPw-pxhqaUOWRFGramr4B1',
+                '_hash': a1Hash,
               },
             },
-            '@layerB': {
-              'nmejjLAUhygiT6WFDPPsHy': {
+            '@tableB': {
+              b0Hash: {
                 'keyB0': 'b0',
-                '_hash': 'nmejjLAUhygiT6WFDPPsHy',
+                '_hash': b0Hash,
               },
-              'dXhIygNwNMVPEqFbsFJkn6': {
+              b1Hash: {
                 'keyB1': 'b1',
-                '_hash': 'dXhIygNwNMVPEqFbsFJkn6',
+                '_hash': b1Hash,
               },
             },
           });
@@ -448,7 +498,7 @@ void main() {
 
         test('with added data', () {
           final rljson2 = rljson.addData({
-            '@layerC': {
+            '@tableC': {
               '_data': [
                 {'keyC0': 'c0'},
               ],
@@ -456,27 +506,27 @@ void main() {
           });
 
           expect(rljson2.data, {
-            '@layerA': {
-              'KFQrf4mEz0UPmUaFHwH4T6': {
+            '@tableA': {
+              a0Hash: {
                 'keyA0': 'a0',
-                '_hash': 'KFQrf4mEz0UPmUaFHwH4T6',
+                '_hash': a0Hash,
               },
-              'YPw-pxhqaUOWRFGramr4B1': {
+              a1Hash: {
                 'keyA1': 'a1',
-                '_hash': 'YPw-pxhqaUOWRFGramr4B1',
+                '_hash': a1Hash,
               },
             },
-            '@layerB': {
-              'nmejjLAUhygiT6WFDPPsHy': {
+            '@tableB': {
+              b0Hash: {
                 'keyB0': 'b0',
-                '_hash': 'nmejjLAUhygiT6WFDPPsHy',
+                '_hash': b0Hash,
               },
-              'dXhIygNwNMVPEqFbsFJkn6': {
+              b1Hash: {
                 'keyB1': 'b1',
-                '_hash': 'dXhIygNwNMVPEqFbsFJkn6',
+                '_hash': b1Hash,
               },
             },
-            '@layerC': {
+            '@tableC': {
               'afNjjrfH8-OfkkEH1uCK14': {
                 'keyC0': 'c0',
                 '_hash': 'afNjjrfH8-OfkkEH1uCK14',